refactor: 💡 refactor the tree structure
