+++ src/commands/badgeedit/interactions.ts
@@ -1,0 +1,23 @@
+import {
+  AutocompleteInteraction,
+  ApplicationCommandOptionChoiceData,
+} from "discord.js";
+
+import { getBadgeChoices } from "commands/_utils/badge";
+
+/**
+ * Handles autocompletion for the badge edit commands
+ * @param interaction
+ * @returns void
+ */
+export const badgeEditAutocomplete = async (
+  interaction: AutocompleteInteraction,
+): Promise<void> => {
+  if (!interaction.isAutocomplete()) return;
+
+  let choices: ApplicationCommandOptionChoiceData[] = [];
+  const focused = interaction.options.getFocused(true);
+  if (focused.name === "badge") choices = getBadgeChoices(focused.value);
+
+  await interaction.respond(choices);
+};
