+++ src/commands/meme/logic.ts
@@ -1,0 +1,13 @@
+import { DiscordInteraction } from "arcybot";
+
+import { getRandomMemeFromAPI } from "api";
+
+/**
+ * Sends a meme to the channel.
+ * @param interaction DiscordInteraction
+ * @returns void
+ */
+export const meme = async (interaction: DiscordInteraction): Promise<void> => {
+  const randomMeme = await getRandomMemeFromAPI();
+  interaction.reply(`_ ${randomMeme} _`);
+};
