+++ src/commands/memeadd/builder.ts
@@ -1,0 +1,13 @@
+import { SlashCommandBuilder } from "discord.js";
+
+/**
+ * Builder for the memeadd command.
+ */
+export const memeaddBuilder = new SlashCommandBuilder()
+  .setName("memeadd")
+  .addStringOption(option =>
+    option
+      .setName("meme")
+      .setDescription("Text or link of the meme")
+      .setRequired(true),
+  );
