+++ src/commands/memedelete/builder.ts
@@ -1,0 +1,13 @@
+import { SlashCommandBuilder } from "discord.js";
+
+/**
+ * Builder for the memedelete command.
+ */
+export const memedeleteBuilder = new SlashCommandBuilder()
+  .setName("memedelete")
+  .addStringOption(option =>
+    option
+      .setName("memeindex")
+      .setDescription("Number of the meme to be deleted")
+      .setRequired(true),
+  );
