+++ src/commands/profile/logic.ts
