+++ src/commands/register/builder.ts
