+++ src/commands/update/logic.ts
