+++ src/commands/vid/builder.ts
