+++ src/commands/vid/logic.ts
