+++ src/consts.ts
@@ -2,3 +2,4 @@ export const DISCORD_MAX_MSG_LENGTH = 2000;
 export const API_URL = "http://89.47.165.141:3002/api";
 
 export const UNKNOWN_NAME = "<UNKNOWN NAME>";
+export const REGISTRATION_REVIEW = "REGISTRATION_REVIEW";
